// src/app/interfaces/userdata.ts
// Generated by https://quicktype.io

export interface UserData {
  id: string;
  chineseName: string;
  englishName: string;
  grade: number;
  gender: Gender;
  position: Position;
  email: string;
  view: string;
  Image: File[] | null;
  state: boolean;
  lock: boolean;
  face_tokens: any[];
  url: string;
  binaryImage: string;
}

export enum Gender {
  女 = "女",
  男 = "男",
}

export enum Position {
  其他 = "其他",
  學生 = "學生",
  老師 = "老師",
}